chore: fix proxy tests
